requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Polymarket client
py-clob-client>=0.0.1

# Dashboard (optional)
streamlit>=1.30.0

# Development
pytest>=7.0.0
